Refactor: Rename chatSignals to chat-signals for consistency
Renamed `frontend/lib/chatSignals.ts` to `frontend/lib/chat-signals.ts` and `frontend/lib/chatSignals.test.ts` to `frontend/lib/chat-signals.test.ts` to follow the project's kebab-case naming convention for utility files in `frontend/lib`. Updated all import references in:
- `frontend/lib/chat-signals.test.ts`
- `frontend/lib/neon/feature-signals.ts`
- `frontend/app/api/chat/send/route.ts`
- `frontend/app/api/insight/map/route.ts`

Ran `bun test frontend/lib/chat-signals.test.ts` to verify functionality.
