+++ frontend/lib/chat-signals.ts
